+++ App/src/screens/Club/index.tsx
@@ -9,7 +9,7 @@ import { ClubAvatar } from '../../components/ClubAvatar';
 import { GoHomeErrorBoundary, withWhoopsErrorBoundary } from '../../components/ErrorBoundary';
 import { MEMBER_HEADER_HEIGHT, MEMBER_HEADER_SCROLL_HEIGHT } from '../../components/Profile/Dimensions';
 import { ProfileHeader } from '../../components/Profile/Header';
-import { withCacheInvalidation } from '../../helper/cache/withCacheInvalidation';
+import { isRecordValid } from '../../helper/cache/withCacheInvalidation';
 import { Categories, Logger } from '../../helper/Logger';
 import { I18N } from '../../i18n/translation';
 import { IClubParams } from '../../redux/actions/navigation';
@@ -111,7 +111,7 @@ class ClubQueryWithPreview extends PureComponent<{
                 }}
                 fetchPolicy={this.props.fetchPolicy}
             >
-                {({ client, loading, data, error }) => {
+                {({ client, loading, data, error, refetch }) => {
                     let preview: any = undefined;
 
                     if (loading || error) {
@@ -157,6 +157,12 @@ class ClubQueryWithPreview extends PureComponent<{
                             }));
                     }
 
+                    if (data && data.Club != null) {
+                        if (!isRecordValid("club", data.Club.LastSync)) {
+                            setTimeout(() => refetch());
+                        }
+                    }
+
                     return React.cloneElement(
                         this.props.children, {
                             loading: loading,
@@ -169,7 +175,6 @@ class ClubQueryWithPreview extends PureComponent<{
     }
 }
 
-const ClubQueryWithPreviewAndInvalidation = withCacheInvalidation("clubs",
-    connect(
+const ClubQueryWithPreviewAndInvalidation = connect(
         null, { addSnack }
-    )(ClubQueryWithPreview));
+    )(ClubQueryWithPreview);