+++ Backend/src/graphql/cache/TTLs.ts
@@ -1,0 +1,6 @@
+export const TTLs = {
+    MemberOverview: 60*4,
+    Member: 60*4,
+    Structure: 60*24,
+    FilterContext: 60*1,
+}